requirements.test.txt
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.0.0
coverage==7.6.12
asynctest==0.13.0

pytest-homeassistant-custom-component==0.13.238